+++ Modules/Microsoft365DSC/Examples/Resources/AADEntitlementManagementAccessPackage/1-Create.ps1
@@ -16,23 +16,14 @@ Configuration Example
     {
         AADEntitlementManagementAccessPackage 'myAccessPackage'
         {
-            DisplayName                     = 'General'
-            AccessPackageResourceRoleScopes = @(
-                MSFT_AccessPackageResourceRoleScope {
-                    Id                                   = 'e5b0c702-b949-4310-953e-2a51790722b8'
-                    AccessPackageResourceOriginId        = '8721d9fd-c6ef-46df-b1b2-bb6f818bce5b'
-                    AccessPackageResourceRoleDisplayName = 'AccessPackageRole'
-                }
-            )
-            CatalogId                       = '1b0e5aca-83e4-447b-84a8-3d8cffb4a331'
-            Description                     = 'Entitlement Access Package Example'
-            IsHidden                        = $false
-            IsRoleScopesVisible             = $true
-            IncompatibleAccessPackages      = @()
-            AccessPackagesIncompatibleWith  = @()
-            IncompatibleGroups              = @()
-            Ensure                          = 'Present'
-            Credential                      = $Credscredential
+            AccessPackagesIncompatibleWith = @();
+            CatalogId                      = "General";
+            Credential                     = $Credscredential;
+            Description                    = "Integration Tests";
+            DisplayName                    = "Integration Package";
+            Ensure                         = "Present";
+            IsHidden                       = $False;
+            IsRoleScopesVisible            = $True;
         }
     }
 }